ab#8365 checkpoint created forecast data service for plots and got plots appearing in the forecast details modal
