+++ projects/common/src/lib/controls/forecast-data.service.spec.ts
@@ -1,0 +1,12 @@
+import { TestBed } from '@angular/core/testing';
+
+import { ForecastDataService } from './forecast-data.service';
+
+describe('ForecastDataServiceService', () => {
+  beforeEach(() => TestBed.configureTestingModule({}));
+
+  it('should be created', () => {
+    const service: ForecastDataService = TestBed.get(ForecastDataService);
+    expect(service).toBeTruthy();
+  });
+});
