+++ projects/common/src/lib/controls/forecast-details/forecast-details.component.html
@@ -15,8 +15,66 @@
     </mat-button-toggle>
   </mat-button-toggle-group>
   <button class="json-download-button" mat-raised-button (click)="ExportToJson()">Download as JSON</button>
-  <div *ngFor="let config of PlotConfigs">
-    <lcu-variable-data-plot [valid-times]="ValidTimes" [forecast-data]="ForecastData" [plot]="config">
-    </lcu-variable-data-plot>
+  <div *ngFor="let chart of Charts">
+    <!-- <lcu-variable-data-plot [valid-times]="ValidTimes" [forecast-data]="ForecastData" [plot]="config">
+    </lcu-variable-data-plot> -->
+    <lcu-charts-line-chart-simple
+  class="chart-container"
+  [animations]="animations"
+  [autoScale]="autoScale"
+  [backgroundGradientConfigs]="chart.chartGradient"
+  [curve]="curve"
+  [gradient]="gradient"
+  [legendPosition]="legendPosition"
+  [legendTitle]="legendTitle"
+  [legend]="showLegend"
+  [maxXAxisTickLength]="maxXAxisTickLength"
+  [maxYAxisTickLength]="maxYAxisTickLength"
+  [rangeFillOpacity]="rangeFillOpacity"
+  [results]="chart.displayData"
+  [rotateXAxisTicks]="rotateXAxisTicks"
+  [roundDomains]="roundDomains"
+  [schemeType]="schemeType"
+  [scheme]="colorScheme"
+  [showGridLines]="showGridLines"
+  [showPercentage]="showPercentage"
+  [showXAxisLabel]="showXAxisLabel"
+  [showYAxisLabel]="showYAxisLabel"
+  [timeline]="timeline"
+  [tooltipDisabled]="tooltipDisabled"
+  [trimXAxisTicks]="trimXAxisTicks"
+  [trimYAxisTicks]="trimYAxisTicks"
+  [units]="chart.yUnits"
+  [view]="view"
+  [xAxisDateFormat]="xAxisDateFormat"
+  [xAxisIsDate]="xAxisIsDate"
+  [xAxisLabel]="xAxisLabel"
+  [xAxis]="showXAxis"
+  [xScaleMax]="xScaleMax"
+  [xScaleMin]="xScaleMin"
+  [yAxisLabel]="yAxisLabel"
+  [yAxisTickFormatting]="chart.YtickFormat"
+  [yAxisTicks]="chart.yAxisTicks"
+  [yAxis]="showYAxis"
+  [dims]="dims"
+  [yScaleMax]="chart.yScaleMax"
+  [yScaleMin]="chart.yScaleMin"
+  [forceManualXHover]="ManualHover"
+  [formatTooltip]="chart.formatTooltip"
+  (activate)="activate($event)"
+  (deactivate)="deactivate($event)"
+  (HoveredVerticalChange)="hoveredVerticalChange($event)"
+  (legendLabelClick)="onLegendLabelClick($event)"
+  (select)="select($event)"
+  >
+  <ng-template #tooltipTemplate let-model="model">
+    <div class="tooltip-item" *ngIf="!chart.formatTooltip">
+        {{model.series}}: {{model.value}}{{chart.yUnits}}
+    </div>
+    <div class="tooltip-item" *ngIf="chart.formatTooltip">
+      {{model.series}}: {{chart.tooltipFormatting(model.value)}}
+  </div>
+  </ng-template>
+</lcu-charts-line-chart-simple>
   </div>
 </mat-dialog-content>
